fix: CategoryRepositoryImpl.create
CategoryRepositoryImpl.create 에서 auto generated key 를 가져올 때
long 이 아닌 int 로 가져오도록 변경

+++ lms/src/main/java/com/lms/repository/category/CategoryRepositoryImpl.java
@@ -62,9 +62,9 @@ public class CategoryRepositoryImpl implements CategoryRepository {
                 Category created = null;
                 try (ResultSet rs = pstmt.getGeneratedKeys()) {
                     if (rs.next()) {
-                        long id = rs.getLong(1);
+                        int id = rs.getInt(1);
                         created = Category.rebuild(
-                                (int) id,
+                                id,
                                 category.getName(),
                                 category.getLevel(),
                                 category.getParentId()
